LocalDispatcher couldn't detect webrick because it hadn't loaded by the time this Engine runs, so I moved it into an :after block.
This allows webrick to be detected.

Also, the base_urls are being cached on SwaggerYard init, so they were ignoring
SwaggerYard.config block in the Rails app initializer.  Setting config.reload is
one way to solve this, but another way is to simply clear the cache after engine
initialization.

since engine is marked :after, config has already occured in the Rails.app, so
clearing the cache forces a reload of the correct baseurl.